fix(restructuring): restructer the files to be well organized
