+++ internal/models/user.go
@@ -1,0 +1,20 @@
+package models
+
+import "time"
+
+// User model
+type User struct {
+	ID        uint      `gorm:"primary_key" json:"id"`
+	Name      string    `json:"name"`
+	Email     string    `json:"email" gorm:"unique"`
+	Password  string    `json:"password"`
+	Username  string    `json:"username" gorm:"unique"`
+	CreatedAt time.Time `json:"created_at"`
+	UpdatedAt time.Time `json:"updated_at"`
+}
+
+// UserProfile for templates
+type UserProfile struct {
+	Name   string
+	Abbrev string
+}
