+++ internal/services/authenticationService/auth_interface.go
@@ -1,0 +1,5 @@
+package authentication
+
+type Authentication_interface interface{
+	
+}
